Rename openfermion/third_party/ to openfermion/contrib
