+++ src/openfermion/contrib/representability/_bijections.py
