+++ src/openfermion/contrib/representability/_dualbasis.py
