+++ src/openfermion/contrib/representability/_higham.py
