+++ src/openfermion/contrib/representability/_higham_test.py
@@ -31,7 +31,7 @@
 from itertools import product
 import numpy as np
 import pytest
-from openfermion.third_party.representability._higham import (
+from openfermion.contrib.representability._higham import (
     heaviside, higham_polynomial, higham_root, map_to_tensor, map_to_matrix,
     fixed_trace_positive_projection)
